requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
antiorm==1.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backpack==0.1
better-profanity==0.7.0
blinker==1.5
braceexpand==0.1.7
charset-normalizer==2.1.1
cleo==0.6.8
click==8.1.3
db==0.1.1
discord-pretty-help==2.0.0
discord.py==2.1.0
Faker==0.8.18
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
inflection==0.3.1
jishaku==2.5.1
jthon==1.0.5
lazy-object-proxy==1.8.0
line-profiler==4.0.1
multidict==6.0.3
mysql-connector-repackaged==0.3.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psutil==5.9.4
psycopg2-binary==2.9.5
pyaml==16.12.2
Pygments==2.13.0
pylev==1.4.0
python-dateutil==2.8.2
pytz==2022.6
pytzdata==2020.1
PyYAML==5.4.1
simplejson==3.18.0
six==1.16.0
text-unidecode==1.2
toml==0.10.2
tzlocal==1.5.1
wrapt==1.14.1
yarl==1.8.2
